Changes place of functions
Also adds execution command to save CSV.
Only saves JSON if dict has length.
All new code was provided by Philip.
